% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes a specified resource group}
\usage{
delete_group(
  GroupName = "string"
)
}
\arguments{
\item{GroupName}{[required] The name of the resource group to delete.}
}
\description{
Deletes a specified resource group. Deleting a resource group does not delete resources that are members of the group; it only deletes the group structure.
}
\examples{

}
