% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{get_group}
\alias{get_group}
\title{Returns information about a specified resource group}
\usage{
get_group(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the resource group.}
}
\description{
Returns information about a specified resource group.
}
\section{Accepted Parameters}{
\preformatted{get_group(
  GroupName = "string"
)
}
}

