% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{get_group_query}
\alias{get_group_query}
\title{Returns the resource query associated with the specified resource group}
\usage{
get_group_query(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the resource group.}
}
\description{
Returns the resource query associated with the specified resource group.
}
\section{Accepted Parameters}{
\preformatted{get_group_query(
  GroupName = "string"
)
}
}

