% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{get_tags}
\alias{get_tags}
\title{Returns a list of tags that are associated with a resource, specified by an ARN}
\usage{
get_tags(Arn)
}
\arguments{
\item{Arn}{The ARN of the resource for which you want a list of tags. The resource must exist within the account you are using.}
}
\description{
Returns a list of tags that are associated with a resource, specified by an ARN.
}
\examples{

}
