% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{list_groups}
\alias{list_groups}
\title{Returns a list of existing resource groups in your account}
\usage{
list_groups(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of resource group results that are returned by ListGroups in paginated output. By default, this number is 50.}

\item{NextToken}{The NextToken value that is returned in a paginated \code{ListGroups} request. To get the next page of results, run the call again, add the NextToken parameter, and specify the NextToken value.}
}
\description{
Returns a list of existing resource groups in your account.
}
\examples{

}
