% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{untag}
\alias{untag}
\title{Deletes specified tags from a specified resource}
\usage{
untag(
  Arn = "string",
  Keys = list(
    "string"
  )
)
}
\arguments{
\item{Arn}{[required] The ARN of the resource from which to remove tags.}

\item{Keys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a specified resource.
}
\examples{

}
