% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{update_group}
\alias{update_group}
\title{Updates an existing group with a new or changed description}
\usage{
update_group(GroupName, Description = NULL)
}
\arguments{
\item{GroupName}{The name of the resource group for which you want to update its description.}

\item{Description}{The description of the resource group. Descriptions can have a maximum of 511 characters, including letters, numbers, hyphens, underscores, punctuation, and spaces.}
}
\description{
Updates an existing group with a new or changed description. You cannot update the name of a resource group.
}
\examples{

}
