% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{create_group}
\alias{create_group}
\title{Creates a group with a specified name, description, and resource query}
\usage{
create_group(Name, ResourceQuery, Description = NULL, Tags = NULL)
}
\arguments{
\item{Name}{The name of the group, which is the identifier of the group in other operations. A resource group name cannot be updated after it is created. A resource group name can have a maximum of 128 characters, including letters, numbers, hyphens, dots, and underscores. The name cannot start with \code{AWS} or \code{aws}; these are reserved. A resource group name must be unique within your account.}

\item{ResourceQuery}{The resource query that determines which AWS resources are members of this group.}

\item{Description}{The description of the resource group. Descriptions can have a maximum of 511 characters, including letters, numbers, hyphens, underscores, punctuation, and spaces.}

\item{Tags}{The tags to add to the group. A tag is a string-to-string map of key-value pairs. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.}
}
\description{
Creates a group with a specified name, description, and resource query.
}
\examples{

}
