% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroups_operations.R
\name{update_group_query}
\alias{update_group_query}
\title{Updates the resource query of a group}
\usage{
update_group_query(GroupName, ResourceQuery)
}
\arguments{
\item{GroupName}{The name of the resource group for which you want to edit the query.}

\item{ResourceQuery}{The resource query that determines which AWS resources are members of the resource group.}
}
\description{
Updates the resource query of a group.
}
\examples{

}
